requirements.txt
asgiref==3.5.2
autopep8==1.6.0
backports.zoneinfo==0.2.1
coverage==6.4
Django==4.0.4
djangorestframework==3.13.1
Pillow==9.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
